Import code from CSE 231.
This class was on computer architecture.  The project is a PDP-11 emulator, written in assembly on the Sparc workstations we had in the university's lab.  The test programs were fed in as octal strings which we had to interpret as machine code and run.  The main test program would reverse an inupt string.